+++ tests/unit/observer/test_flaky_test_aggregator.py
@@ -57,8 +57,9 @@ class TestFlakyTestAggregator:
         assert len(result.flaky_tests) > 0
         assert result.flaky_tests[0]["test_name"] == "tests/test_foo.py::test_flaky"
 
-    @pytest.mark.skip(
-        reason="Test has logic bug: expects sum of session counts but gets single session value"
+    @pytest.mark.xfail(
+        strict=False,
+        reason="Test has logic bug: expects sum of session counts but gets single session value",
     )
     def test_aggregate_multiple_sessions(self, tmp_path):
         """Test aggregation across multiple sessions."""
